Move mutmap to separate repository
The goal of this commit is to move the mutmap (visualization)
code into a separate repository, without losing any of the commit
history. I did this by removing everything except the mutmap
directory.

This might not be the best way to do this, but no information is
lost, so a better approach could be used in the future.